doc: Add Roadmap to README (#47)
## Problem

In my first reading of `hermes-mcp`, I couldn't tell if it was meant to
be a MCP client, or both MCP client and server. (turns out it is the
latter) I hope this PR helps future readers avoid the same confusion.

## Solution

In the README, I change `## Features` to `## Roadmap`, using checkboxes
to indicate what is done and what is planned.

## Rationale

I took a guess at a set of planned features.  Edit as you see fit.  ;-)