config: multi version pipeline (#192)
This pull request updates the CI configuration in
`.github/workflows/ci.yml` to expand the matrix of Elixir and OTP
versions for testing. The changes ensure broader compatibility by
testing multiple combinations of Elixir and OTP versions.

Updates to CI matrix configuration:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL20-R30):
Replaced the single combination of `elixir: 1.19.0-rc.0` and `otp: 27.3`
with multiple combinations, including `elixir: 1.19` with `otp: 28` and
`27`, and `elixir: 1.18` with `otp: 28`, `27`, and `26`. This change was
repeated across three sections of the file.
[[1]](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL20-R30)
[[2]](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL75-R94)
[[3]](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fL143-R171)